// src/interfaces.ts
export interface GameAudio {
    correct: HTMLAudioElement;
    victory: HTMLAudioElement;
    gameOver: HTMLAudioElement;
    wrongAnswer: HTMLAudioElement;
};

export interface HangmanElements {
    ground: HTMLElement;
    scaffold: HTMLElement;
    head: HTMLElement;
    body: HTMLElement;
    arms: HTMLElement;
    legs: HTMLElement;
    hearts: HTMLElement;
};